+++ app/src/main/java/br/com/bg7/appvistoria/data/source/remote/http/HttpProgressCallback.java
@@ -5,5 +5,5 @@ package br.com.bg7.appvistoria.data.source.remote.http;
  * Date: 2017-07-28
  */
 public interface HttpProgressCallback<T> extends HttpCallback<T>{
-    void onProgressUpdated(double percentage);
+    void onProgressUpdated(Integer percentage);
 }
